third_party/blink/web_tests/wpt_internal/html/canvas/drawElementImage/privacy/svg-images-ignored.https.sub.html

<!DOCTYPE HTML>
<html>
<head>
  <title>drawElementImage does not draw cross-origin SVG images</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
</head>
<body>
  <canvas id=canvas width="100" height="200" layoutsubtree>
    <svg id=child width=100 height=200>
      <rect fill="green" x=0 y=0 width=100 height=200 />
      <image id="sameOrigin" href="https://{{location[host]}}/wpt_internal/html/canvas/drawElementImage/resources/red-100x100.png" x=10 y=10 height=100 width=100 />
      <image id="crossOrigin" href="https://{{hosts[alt][www]}}:{{ports[h2][0]}}/wpt_internal/html/canvas/drawElementImage/resources/red-100x100.png" x=0 y=100 height=100 width=100 />
    </svg>
  </canvas>

  <script>
    window.onload = function() {
      test(function(t) {
        var context = canvas.getContext("2d");
        context.drawElementImage(child, 0, 0);

        pixel = context.getImageData(50, 50, 1, 1).data;
        assert_equals(pixel[0], 255, "Same origin should draw");
        assert_equals(pixel[1], 0, "Same origin should draw");
        assert_equals(pixel[2], 0, "Same origin should draw");

        pixel = context.getImageData(50, 150, 1, 1).data;
        assert_equals(pixel[0], 0, "Cross origin should not draw");
        assert_equals(pixel[1], 128, "Cross origin should not draw");
        assert_equals(pixel[2], 0, "Cross origin should not draw");

        t.done();
      });
    };
  </script>
</body>
</html>
